Make sig_extra in TcSigInfo a (Maybe SrcSpan)
Previously, sig_extra was a Boolean indicating the presence of an
extra-constraints wildcard. Now we store a (Maybe SrcSpan). We store the
location iff an extra-constraints wildcard is present. We'll need this SrcSpan
later to have an accurate location when we inform the user of the extra
constraints the wildcard generated.