Delayed reporting of the wildcard instantiations
The final types the wildcards in partial type signatures were instantiated to
are only known after simplifyTop is called, at which stage we no longer have
the necessary information to report the instantiations (the locations of the
wildcards, etc.). Therefore, save the reporters as delayed monadic
computations in the environment when we have the required information, and
retrieve and execute them after the final types are known.