Deal with the monomorphism restriction for pts
Regarding the monomorphism restriction, we treat a binding with a partial type
signature like a binding without a type signature. However, type variables
mentioned in the partial type signature are considered unrestricted. This
granularity wasn't implemented before, as there was no need, there only used
to be: all type variable are mentioned in the type signature or none at all.